flock_resources/pyproject.toml
[tool.poetry]
name = "flock-resources"
version = "0.0.1"
description = ""
authors = ["Adam Yodinsky <27074934+adamyodinsky@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "builder", from = "src" },
    { include = "resources", from = "src" },
    { include = "schemas", from = "src" },
    { include = "store", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.7"
pyyaml = "^6.0"
langchain = "^0.0.191"
chromadb = "^0.3.21"
tiktoken = "^0.3.3"
google-search-results = "^2.4.2"
openai = "^0.27.4"
faiss-cpu = "^1.7.3"
flock-common = {path = "../flock_common", develop = true}
odmantic = "^0.9.2"
python-dotenv = "^1.0.0"
wolframalpha = "^5.0.0"
pdfminer-six = "^20221105"
azure-core = "^1.26.4"
beautifulsoup4 = "^4.12.2"
pytest-playwright = "^0.3.3"
pymongo = "^4.3.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
